AISJapan系の処理を別プロジェクトに分離した
